requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
autoflake==1.4
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
click==8.1.2
cycler==0.11.0
fastapi==0.75.2
flatbuffers==2.0
fonttools==4.32.0
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
onnxruntime==1.11.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pip==21.2.4
protobuf==3.20.0
psutil==5.9.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-lzf==0.2.4
python-multipart==0.0.5
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing==3.7.4.3
typing_extensions==4.2.0
uvicorn==0.17.6
wheel==0.37.1
